Add accumulated list of nieghbourCells
Add a function to stack all occurrences of neighbour cells into a list. Tricky to do this without modifying a counter, this is the  first idea I have for how to count how many neighbours a cell has. Will complete it with function that counts occurrences of a single cell in this list. This will then be how many neighbours that cell has, This is the input for the main logic of the game.